requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.1
pgvector==0.3.6
openai==1.40.6